fix: move versions.json back to root
